Dissolve the T5MainLayer class
This also removes support for key/value caching and gated activation
that we are not current using in any forward pass (we can always add
these back).
