+++ keras_nlp/models/t5/t5_presets.py
@@ -23,7 +23,6 @@ backbone_presets = {
             "model_card": "https://github.com/google-research/text-to-text-transfer-transformer/blob/main/released_checkpoints.md",
         },
         "config": {
-            "use_gated_activation": False,
             "activation": "relu",
             "intermediate_dim": 2048,
             "hidden_dim": 512,
@@ -50,7 +49,6 @@ backbone_presets = {
             "model_card": "https://github.com/google-research/text-to-text-transfer-transformer/blob/main/released_checkpoints.md",
         },
         "config": {
-            "use_gated_activation": False,
             "activation": "relu",
             "intermediate_dim": 3072,
             "hidden_dim": 768,
@@ -77,7 +75,6 @@ backbone_presets = {
             "model_card": "https://github.com/google-research/text-to-text-transfer-transformer/blob/main/released_checkpoints.md",
         },
         "config": {
-            "use_gated_activation": False,
             "activation": "relu",
             "intermediate_dim": 4096,
             "hidden_dim": 1024,
@@ -104,7 +101,6 @@ backbone_presets = {
             "model_card": "https://github.com/google-research/text-to-text-transfer-transformer/blob/main/released_checkpoints.md",
         },
         "config": {
-            "use_gated_activation": False,
             "activation": "relu",
             "intermediate_dim": 16384,
             "hidden_dim": 1024,
